+++ test/ZendRouterTest.php
@@ -1,14 +1,13 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       http://github.com/zendframework/zend-diactoros for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-diactoros/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/zendframework/zend-expressive-zendrouter for the canonical source repository
+ * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-zendrouter/blob/master/LICENSE.md New BSD License
  */
 
 namespace ZendTest\Expressive\Router;
 
+use Fig\Http\Message\RequestMethodInterface as RequestMethod;
 use PHPUnit_Framework_TestCase as TestCase;
 use Prophecy\Argument;
 use Zend\Diactoros\ServerRequest;
@@ -37,7 +36,7 @@ class ZendRouterTest extends TestCase
         $this->assertAttributeInstanceOf(TreeRouteStack::class, 'zendRouter', $router);
     }
 
-    public function createRequestProphecy()
+    public function createRequestProphecy($requestMethod = RequestMethod::METHOD_GET)
     {
         $request = $this->prophesize('Psr\Http\Message\ServerRequestInterface');
 
@@ -45,8 +44,8 @@ class ZendRouterTest extends TestCase
         $uri->getPath()->willReturn('/foo');
         $uri->__toString()->willReturn('http://www.example.com/foo');
 
-        $request->getMethod()->willReturn('GET');
-        $request->getUri()->willReturn($uri);
+        $request->getMethod()->willReturn($requestMethod);
+        $request->getUri()->will([$uri, 'reveal']);
         $request->getHeaders()->willReturn([]);
         $request->getCookieParams()->willReturn([]);
         $request->getQueryParams()->willReturn([]);
@@ -80,7 +79,7 @@ class ZendRouterTest extends TestCase
                 'GET' => [
                     'type' => 'method',
                     'options' => [
-                        'verb' => 'GET',
+                        'verb' => 'GET,HEAD,OPTIONS',
                         'defaults' => [
                             'middleware' => 'foo',
                         ],
@@ -126,7 +125,7 @@ class ZendRouterTest extends TestCase
                 'GET' => [
                     'type' => 'method',
                     'options' => [
-                        'verb' => 'GET',
+                        'verb' => 'GET,HEAD,OPTIONS',
                         'defaults' => [
                             'middleware' => 'foo',
                         ],
@@ -188,7 +187,7 @@ class ZendRouterTest extends TestCase
                 'GET' => [
                     'type' => 'method',
                     'options' => [
-                        'verb' => 'GET',
+                        'verb' => 'GET,HEAD,OPTIONS',
                         'defaults' => [
                             'middleware' => 'foo',
                         ],
@@ -269,10 +268,14 @@ class ZendRouterTest extends TestCase
         $this->zendRouter
             ->match(Argument::type(ZendRequest::class))
             ->willReturn($routeMatch->reveal());
+        $this->zendRouter
+            ->addRoute('/foo', Argument::type('array'))
+            ->shouldBeCalled();
 
         $request = $this->createRequestProphecy();
 
         $router = $this->getRouter();
+        $router->addRoute(new Route('/foo', 'bar', [RequestMethod::METHOD_GET], '/foo'));
         $result = $router->match($request->reveal());
         $this->assertInstanceOf(RouteResult::class, $result);
         $this->assertTrue($result->isSuccess());
@@ -365,5 +368,60 @@ class ZendRouterTest extends TestCase
         $result = $router->match($request);
         $this->assertTrue($result->isFailure());
         $this->assertFalse($result->isMethodFailure());
+    }
+
+    public function testSuccessfulMatchingComposesRouteInRouteResult()
+    {
+        $route = new Route('/foo', 'bar', [RequestMethod::METHOD_GET]);
+
+        $routeMatch = $this->prophesize(RouteMatch::class);
+        $routeMatch->getMatchedRouteName()->willReturn($route->getName());
+        $routeMatch->getParams()->willReturn([
+            'middleware' => $route->getMiddleware(),
+        ]);
+
+        $this->zendRouter
+            ->match(Argument::type(ZendRequest::class))
+            ->willReturn($routeMatch->reveal());
+        $this->zendRouter
+            ->addRoute('/foo^GET', Argument::type('array'))
+            ->shouldBeCalled();
+
+        $request = $this->createRequestProphecy();
+
+        $router = $this->getRouter();
+        $router->addRoute($route);
+
+        $result = $router->match($request->reveal());
+
+        $this->assertInstanceOf(RouteResult::class, $result);
+        $this->assertTrue($result->isSuccess());
+        $this->assertSame($route, $result->getMatchedRoute());
+    }
+
+    public function implicitMethods()
+    {
+        return [
+            'head'    => [RequestMethod::METHOD_HEAD],
+            'options' => [RequestMethod::METHOD_OPTIONS],
+        ];
+    }
+
+    /**
+     * @dataProvider implicitMethods
+     */
+    public function testRoutesCanMatchImplicitHeadAndOptionsRequests($method)
+    {
+        $route = new Route('/foo', 'bar', [RequestMethod::METHOD_PUT]);
+
+        $router = new ZendRouter();
+        $router->addRoute($route);
+
+        $request = $this->createRequestProphecy($method);
+        $result = $router->match($request->reveal());
+
+        $this->assertInstanceOf(RouteResult::class, $result);
+        $this->assertTrue($result->isSuccess());
+        $this->assertSame($route, $result->getMatchedRoute());
     }
 }